% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcens.R
\name{sourcens}
\alias{sourcens}
\title{Load a file into a corresponding namespace.}
\usage{
sourcens(path)
}
\arguments{
\item{path}{the path of the file to be parsed}
}
\description{
Load a file into a corresponding namespace.
Errors are shown but execution is continued.
usage: sourceNs("filename.R") - sources file into namespace  'filename::'
}
